Fix/changing doubles in components #5

+++ src/main/java/simpleerp/component/ComponentResponseDTO.java
@@ -4,6 +4,7 @@ package simpleerp.component;
 import simpleerp.common.Unit;
 import simpleerp.vendor.ShowComponentVendorDTO;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 
@@ -11,7 +12,7 @@ public record ComponentResponseDTO(
         Long id,
         String name,
         Unit unit,
-        Double costPerUnit,
+        BigDecimal costPerUnit,
         List<ShowComponentVendorDTO> componentVendorList
 ) {
 }
